Standardize step definition use of transforms. Closes #9
