+++ lib/step_definitions/GeneralStepDefs.js
@@ -40,8 +40,12 @@ defineSupportCode(function({Given, When, Then, setWorldConstructor}) {
     return this.currentPage[list].element(protractor.By.cssContainingText('option', optionText)).click();
   });
 
-  Then('the title should equal "{text:captureString}"', function(text) {
-    return this.expect(browser.getTitle()).to.eventually.equal(text);
+  Then('the title {expectation:shouldToBoolean} equal "{text:captureString}"', function(expectation, text) {
+    if (expectation) {
+      return this.expect(browser.getTitle()).to.eventually.equal(text);
+    } else {
+      return this.expect(browser.getTitle()).not.to.eventually.equal(text);
+    }
   });
 
   Then('the "{name:elementName}"{type:elementType} {expectation:shouldToBoolean} be active', function(name, type, expectation) {
@@ -49,9 +53,9 @@ defineSupportCode(function({Given, When, Then, setWorldConstructor}) {
     return this.expect(this.elementHelper.hasClass(this.currentPage[element], 'active')).to.eventually.equal(expectation);
   });
 
-  Then('the "{name:elementName}"{type:elementType} should be present', function(name, type) {
+  Then('the "{name:elementName}"{type:elementType} {expectation:shouldToBoolean} be present', function(name, type, expectation) {
     var element = name + type;
-    return this.expect(this.currentPage[element].isPresent()).to.eventually.equal(true);
+    return this.expect(this.currentPage[element].isPresent()).to.eventually.equal(expectation);
   });
 
   Then('I (should be on|reach|am taken to) the "{pageName:captureString}" page', function(pageName) {
@@ -59,7 +63,7 @@ defineSupportCode(function({Given, When, Then, setWorldConstructor}) {
     return this.currentPage.waitForLoaded();
   });
 
-  Then('(the |)"{name:elementName}"{type:elementType} should (have|contain) the text "{text:captureString}"', function(name, type, text) {
+  Then('(the |)"{name:elementName}"{type:elementType} {expectation:shouldToBoolean} (have|contain) the text "{text:captureString}"', function(name, type, expectation, text) {
     this.element = this.currentPage[name + type];
 
     var elementText = this.element.getTagName().then((tagName) => {
@@ -70,17 +74,25 @@ defineSupportCode(function({Given, When, Then, setWorldConstructor}) {
       };
     });
 
-    return this.expect(elementText).to.eventually.contain(text);
+    if (expectation) {
+      return this.expect(elementText).to.eventually.contain(text);
+    } else {
+      return this.expect(elementText).not.to.eventually.contain(text);
+    }
   });
 
-  Then('"{option:captureString}" should appear in the "{name:elementName}"{type:elementType}', function(option, name, type) {
+  Then('"{option:captureString}" {expectation:shouldToBoolean} appear in the "{name:elementName}"{type:elementType}', function(option, expectation, name, type) {
     this.list = this.currentPage[name + type];
 
     var optionsText = this.list.all(By.tagName('option')).map(function(element, index) {
       return element.getText();
     });
 
-    return this.expect(optionsText).to.eventually.contain(option);
+    if (expectation) {
+      return this.expect(optionsText).to.eventually.contain(option);
+    } else {
+      return this.expect(optionsText).not.to.eventually.contain(option);
+    }
   });
 
   Then('the "{name:elementName}"{type:elementType} {expectation:shouldToBoolean} be displayed', function(name, type, expectation) {
