+++ spec/generalStepDefs/the_title_should_equal.spec.js
@@ -10,28 +10,48 @@ describe('the title should equal "___"', function() {
     });
 
     before(function() {
-      stepPattern = 'the title should equal "{text:captureString}"';
+      stepPattern = 'the title {expectation:shouldToBoolean} equal "{text:captureString}"';
     });
 
-    it('should match \'the title should equal "My Title"\'', function() {
+    it('should match "...should equal..."', function() {
       verifyStepMatch('the title should equal "My Title"');
     });
 
-    it('should capture the title', function() {
-      verifyStepCaptures('the title should equal "My Title"', 'My Title');
+    it('should match "...should not equal..."', function() {
+      verifyStepMatch('the title should not equal "My Title"');
+    });
+
+    it('should capture the title and the expectation', function() {
+      verifyStepCaptures('the title should equal "My Title"', 'should', 'My Title');
     });
   });
 
   describe('execution', function() {
-    it('should succeed if the page title matches the supplied title', function() {
-      return executeStep('the title should equal "Protractor Integration Test Page"', function() {
-        expect(currentStepResult.status).to.equal(Cucumber.Status.PASSED);
+    describe('with a "should" expectation', function() {
+      it('should succeed if the page title matches the supplied title', function() {
+        return executeStep('the title should equal "Protractor Integration Test Page"', function() {
+          expect(currentStepResult.status).to.equal(Cucumber.Status.PASSED);
+        });
+      });
+
+      it('should fail if the page title does not match the supplied title', function() {
+        return executeStep('the title should equal "Fake Title"', function() {
+          expect(currentStepResult.status).to.equal(Cucumber.Status.FAILED);
+        });
       });
     });
 
-    it('should fail if the page title does not match the supplied title', function() {
-      return executeStep('the title should equal "Fake Title"', function() {
-        expect(currentStepResult.status).to.equal(Cucumber.Status.FAILED);
+    describe('with a "should not" expectation', function() {
+      it('should fail if the page title matches the supplied title', function() {
+        return executeStep('the title should not equal "Protractor Integration Test Page"', function() {
+          expect(currentStepResult.status).to.equal(Cucumber.Status.FAILED);
+        });
+      });
+
+      it('should succeed if the page title does not match the supplied title', function() {
+        return executeStep('the title should not equal "Fake Title"', function() {
+          expect(currentStepResult.status).to.equal(Cucumber.Status.PASSED);
+        });
       });
     });
   });
